fix bugs from previous commit

+++ src/core/GenerateSDFMaterial.js
@@ -31,9 +31,7 @@ export class GenerateSDFMaterial extends ShaderMaterial {
 				${shaderDistanceFunction}
 				varying vec2 vUv;
         uniform BVH bvh;
-				// uniform float zValue;
-        // change
-        uniform int zValue;
+				uniform float zValue;
 				uniform mat4 matrix;
 				void main() {
           // compute the point in space to check
